+++ src/GrumPHP/Collection/TaskResultCollection.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace GrumPHP\Collection;
+
+use Doctrine\Common\Collections\ArrayCollection;
+
+class TaskResultCollection extends ArrayCollection
+{
+    /**
+     * @return bool
+     */
+    public function isPassed()
+    {
+        return $this->forAll(function ($key, $taskResult) {
+            return $taskResult->isPassed();
+        });
+    }
+}
